hack/benchmark/requirements.txt
matplotlib==3.10.8
numpy==2.3.3
PyYAML==6.0.2
